### Clang
- 为 switch 添加 default 
- int 和 size_t 比较之前先 static_cast
- 添加两处 [[maybe_unused]]
- 修正一处参数列表的顺序

### GCC
- 显式忽略一处 missing field initializers

### CMake
- 由于目前的所有 warning 都修复了，编译参数就添加了 -Werror（希望不会造成后续的不便）

+++ src/MaaCore/Task/Fight/StageNavigationTask.cpp
@@ -49,7 +49,7 @@ bool asst::StageNavigationTask::set_stage_name(const std::string& stage_name)
     if (!difficulty.empty()) {
         std::string upper_difficulty = difficulty;
         upper_difficulty[0] = static_cast<char>(::toupper(upper_difficulty[0]));
-        for (int i = 1; i < upper_difficulty.size(); ++i) {
+        for (size_t i = 1; i < upper_difficulty.size(); ++i) {
             upper_difficulty[i] = static_cast<char>(::tolower(upper_difficulty[i]));
         }
         static const std::string DifficultyTaskPrefix = "ChapterDifficulty";
